e2e, multi-homing: move stuff around to an utils file
This way, the test file only has the test themselves.
